iot-devices/AGRISENSE_IoT/arduino_nano_firmware/requirements.txt
# Arduino Bridge Requirements
# Python dependencies for the Arduino Nano serial bridge

pyserial>=3.5
requests>=2.28.0
python-dotenv>=0.19.0